Split up File plug into 2 classes
